+++ src/bandcamp/inject.js
@@ -2,7 +2,12 @@
  * Injected script into Bandcamp page.
  * We can send any window data to our content.js script.
  */
-window.postMessage({
-  type: 'BANDCAMP_DATA',
-  bandData: window.BandData
-});
+(function () {
+  const pageData = JSON.parse(document.getElementById('pagedata').dataset.blob);
+
+  window.postMessage({
+    type: 'BANDCAMP_DATA',
+    bandData: window.BandData,
+    pageData
+  });
+})();