+++ src/bandcamp/pages/page-music.js
@@ -23,7 +23,7 @@ import { log } from '../../utils/console';
 import { chromeListenToMessage } from '../../utils/chrome';
 import { Music } from '../../app/music';
 import { ArtistItem } from '../../app/artistItem';
-import { Storage } from '../../app/core/storage';
+import { Storage, STORAGE_KEY } from '../../app/core/storage';
 
 const storage = new Storage();
 
@@ -31,10 +31,28 @@ const storage = new Storage();
 export function setupPageMusic(pageType) {
   listenForMessage('BANDCAMP_DATA', (messageData) => {
     const music = createMusic(messageData.bandData);
-    storage.save(music);
+
+    storage.save(music).then(() => {
+      savePageData(messageData.pageData);
+    });
+
     setupSendMessageToPopup(pageType, music);
   });
   setupIsotope();
+}
+
+function savePageData(pageData) {
+  if (!pageData.identities.fan) return;
+
+  const username = pageData.identities.fan.username;
+  const url = pageData.identities.fan.url;
+
+  storage.set(STORAGE_KEY.BANDCAMP_DATA, {
+    user: {
+      username,
+      url
+    }
+  });
 }
 
 function createMusic(bandData) {